requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
requests
cairocffi
svgwrite
pillow
cairosvg
IPython
imageio
rdp
streamlit-drawable-canvas
python-dotenv
colorama
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
